refactor: extract products, orders, revenue, and form routes (#108)
app.js is now wiring only: it builds the admin gate from the auth seam
and mounts one router per area, matching the money-path layout from
#107. Handlers moved verbatim; the shared form rate-limit bucket and
route registration order are preserved. Closes out the #3 slicing.
